requirements.txt
pinecone
python-dotenv
pdfplumber
groq
